Refactor modal set up
