+++ ui/DesignSystem/Component/index.js
@@ -23,7 +23,6 @@ import Rad from "./Rad.svelte";
 import RadicleLogo from "./RadicleLogo.svelte";
 import RadioOption from "./RadioOption.svelte";
 import Remote from "./Remote.svelte";
-import SearchModal from "./SearchModal.svelte";
 import RemoteHelperHint from "./RemoteHelperHint.svelte";
 import SegmentedControl from "./SegmentedControl.svelte";
 import Sidebar from "./Sidebar.svelte";
@@ -73,7 +72,6 @@ export {
   Remote,
   RemoteHelperHint,
   Row,
-  SearchModal,
   SegmentedControl,
   Sidebar,
   SidebarLayout,
