+++ ui/Screen/Modal/Shortcuts.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Illustration from "./Illustration.svelte";
+  import { Illustration } from "../../DesignSystem/Component";
   import { isMac } from "../../src/settings.ts";
   import { Variant as IllustrationVariant } from "../../src/illustration";
