+++ ui/Screen/Modal/index.js
@@ -1,0 +1,9 @@
+import NewProject from "./ProjectCreation.svelte";
+import Search from "./SearchModal.svelte";
+import Shortcuts from "./Shortcuts.svelte";
+
+export default {
+  NewProject,
+  Search,
+  Shortcuts,
+};
