+++ ui/Screen/Profile/ProjectsMenu.svelte
@@ -1,11 +1,11 @@
 <script>
-  import * as path from "../../src/path.ts";
   import * as modal from "../../src/modal.ts";
+  import * as path from "../../src/path.ts";
 
   import { Icon, Button } from "../../DesignSystem/Primitive";
 </script>
 
-<span on:click|stopPropagation={() => modal.toggle(path.createProject())}>
+<span on:click|stopPropagation={() => modal.toggle(path.newProject())}>
   <Button
     dataCy="new-project-button"
     variant="outline"
